22.04 package fine tuning

+++ post_install/docker.sh
@@ -9,24 +9,32 @@ source ./common_bash_funcs.sh
 
 apt_update
 
-echo "attempting to remove old docker packages"
-sudo DEBIAN_FRONTEND=noninteractive apt-get remove -y --purge docker docker-ce docker-ce-cli docker-engine docker-compose docker.io containerd runc >/dev/null 2>&1
-sudo rm -rf /etc/apt/sources.list.d/docker*
+if [[ 1 == 0 ]]; then
+	echo "attempting to remove old docker packages"
+	sudo DEBIAN_FRONTEND=noninteractive apt-get remove -y --purge docker docker-ce docker-ce-cli docker-engine docker-compose docker.io containerd runc >/dev/null 2>&1
+	sudo rm -rf /etc/apt/sources.list.d/docker*
+fi
 
 sudo systemctl status docker --no-pager
 
-curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
+if [[ 1 == 1 ]]; then
+	apt_update
+	apt_group_install_auto_yes "docker.io docker-compose"
+	sudo DEBIAN_FRONTEND=noninteractive apt autoremove -y
+else
+	curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
 
-sudo apt-key fingerprint 0EBFCD88
+	sudo apt-key fingerprint 0EBFCD88
 
-sudo systemctl status docker --no-pager
-curl -fsSL https://get.docker.com -o get-docker.sh
-sudo sh get-docker.sh
-rm -rf get-docker.sh
+	sudo systemctl status docker --no-pager
+	curl -fsSL https://get.docker.com -o get-docker.sh
+	sudo sh get-docker.sh
+	rm -rf get-docker.sh
 
-apt_update
-apt_group_install_auto_yes "docker-ce docker-ce-cli containerd.io docker-compose"
-sudo DEBIAN_FRONTEND=noninteractive apt autoremove -y
+	apt_update
+	apt_group_install_auto_yes "docker-ce docker-ce-cli containerd.io docker-compose"
+	sudo DEBIAN_FRONTEND=noninteractive apt autoremove -y
+fi
 
 sudo groupadd docker
 sudo usermod -aG docker `getent group sudo | awk -F: '{print $4}'`
